updated top populated cities in the region (added function call)

+++ src/main/java/com/napier/coursework/City.java
@@ -8,7 +8,7 @@ public class City {
 
     private String District;
 
-    private double Population;
+    private int Population;
 
     public void setCityName(String cityName) {
         CityName = cityName;
